% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com-registry.R
\name{getInstrument}
\alias{getInstrument}
\title{Retrieve a registered clinical instrument}
\usage{
getInstrument(id)
}
\arguments{
\item{id}{Instrument identifier.}
}
\value{
The \link{ClinicalInstrument}.
}
\description{
Retrieve a registered clinical instrument
}
\seealso{
\code{\link[=listInstruments]{listInstruments()}}, \code{\link[=registerInstrument]{registerInstrument()}}
}
