% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinimetrics-store.R
\name{listClinimetrics}
\alias{listClinimetrics}
\title{List available clinimetric constants}
\usage{
listClinimetrics(instrument = NULL)
}
\arguments{
\item{instrument}{Optional instrument to filter by (case insensitive).}
}
\value{
A \code{data.frame} of the stored clinimetric constants.
}
\description{
List available clinimetric constants
}
\examples{
listClinimetrics("6MWT")
}
\seealso{
\code{\link[=getClinimetric]{getClinimetric()}}
}
