% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com-registry.R
\name{listInstruments}
\alias{listInstruments}
\title{List available clinical instruments}
\usage{
listInstruments()
}
\value{
A character vector of registered instrument ids.
}
\description{
List available clinical instruments
}
\seealso{
\code{\link[=getInstrument]{getInstrument()}}, \code{\link[=registerInstrument]{registerInstrument()}}
}
