% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioClinical-package.R
\docType{package}
\name{PhysioClinical-package}
\alias{PhysioClinical}
\alias{PhysioClinical-package}
\title{PhysioClinical: clinical outcome measures and responder analysis}
\description{
Validated outcome-measure scoring, MDC/MCID responder classification, ICF
tagging, and normative deviation scoring for the x-biosignal ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioClinical}
  \item \url{https://x-biosignal.github.io/PhysioClinical/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioExperiment/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
