% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normative-print.R
\name{show,GovernedNormativeReference-method}
\alias{show,GovernedNormativeReference-method}
\title{Show a normative reference (surfacing its governance)}
\usage{
\S4method{show}{GovernedNormativeReference}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{GovernedNormativeReference}}.}
}
\value{
\code{object}, invisibly.
}
\description{
Show a normative reference (surfacing its governance)
}
